[AOS] refactor: di refactor
